[domain] move Exclude into module
